refactor: switch to TS
