+++ src/index.ts
@@ -1,0 +1,6 @@
+import { httpServer } from './http_server/index.js';
+
+const HTTP_PORT = 8181;
+
+console.log(`Start static http server on the ${HTTP_PORT} port!`);
+httpServer.listen(HTTP_PORT);
